requirements.txt
Django==2.2.5
psycopg2==2.8.3
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.1.0
dj-database-url==0.5.0
django-heroku==0.3.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.4
